feat: add confirmation-dialog

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -6,6 +6,7 @@ import {MatDialog} from "@angular/material/dialog";
 import {ErrorDialogComponent} from "../../../shared/components/error-dialog/error-dialog.component";
 import {ActivatedRoute, Router} from "@angular/router";
 import {MatSnackBar} from "@angular/material/snack-bar";
+import {ConfirmationDialogComponent} from "../../../shared/confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: 'app-courses',
@@ -51,16 +52,24 @@ export class CoursesComponent {
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.refresh()
-        this.snackBar.open('Curso removido com sucesso!', 'X',
-          {duration: 5000,
-            verticalPosition: 'top',
-            horizontalPosition: 'center'})
-      },
-      error => this.openError('Erro ao tentar remover curso.')
-    )
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if(result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.refresh()
+            this.snackBar.open('Curso removido com sucesso!', 'X',
+              {duration: 5000,
+                verticalPosition: 'top',
+                horizontalPosition: 'center'})
+          },
+          error => this.openError('Erro ao tentar remover curso.')
+        )
+      }
+    });
   }
 
   ngOnInit(): void {
